Restructured Files for better clarity
